Replace float values for decimal
Filters don't work with float numbers. If you try to use a filter in a float column it won't work because internally float numbers may not be stored in the same way we insert them. For example: 43.23 can be stored as 43.229999999999999. This replace the float type by the decimal type that works as we expect.